Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/traverse_the_history_1-expected.txt

main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)

FAIL Multiple history traversals from the same task assert_array_equals: Pages opened during history navigation property 1, expected 2 but got 3
